+++ src/com/spz/controller/RolesController.java
@@ -1,5 +1,6 @@
 package com.spz.controller;
 
+import java.util.List;
 import java.util.TreeSet;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,8 +11,10 @@ import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.spz.dao.RolesMapper;
 import com.spz.entity.Roles;
+import com.spz.entity.UserRoles;
 import com.spz.service.RolesModulesService;
 import com.spz.service.RolesService;
+import com.spz.service.UserRolesService;
 import com.spz.util.Result;
 
 @Controller
@@ -21,6 +24,7 @@ public class RolesController {
 	
 	@Autowired private RolesModulesService rolesModulesService;
 	
+	@Autowired private UserRolesService  userRolesService;
 	
 	@RequestMapping(value="/selectAllRoles",method=RequestMethod.POST)
 	@ResponseBody
@@ -44,8 +48,14 @@ public class RolesController {
 	@RequestMapping(value="/deleteRolesById",method=RequestMethod.POST)
 	@ResponseBody
 	public String deleteRolesById(Integer r_id) {
-		Integer roles = rolesService.deleteRoles(r_id);
-		return Result.toClient(roles>0 ? true :false, roles>0 ? "删除成功" : "删除失败");
+		//删除角色之前首先判断是否被员工引用，如果被引用的时候不能删除
+		List<UserRoles> byR_id = userRolesService.selectByR_id(r_id);
+		if(byR_id.size()>0) {
+			return Result.toClient(true, "该角色正在被员工引用，请解除后再进行删除！");
+		}else {
+			Integer roles = rolesService.deleteRoles(r_id);
+			return Result.toClient(roles>0 ? true :false, roles>0 ? "删除成功" : "删除失败");
+		}
 	}
 	
 	@RequestMapping(value="/setupQuanXian",method=RequestMethod.POST)
@@ -53,6 +63,10 @@ public class RolesController {
 	public String setupQuanXian(String arr,Integer r_id) {
 		//拿到角色的id去角色和模块的中间表删除所有角色的模块，然后去添加新的模块信息
 		String[] split = arr.split(",");
+		if(split.length==1) {
+			 Integer modules = rolesModulesService.deleteRoleModules(r_id);
+			 return Result.toClient(true, modules>0 ? "设置成功" : "设置失败");
+		}
 		TreeSet<String> tr = new TreeSet<String>();
 		for (int i = 0; i < split.length; i++) {
 			tr.add(split[i]);
@@ -64,8 +78,8 @@ public class RolesController {
 		Integer integer = rolesModulesService.deleteRoleModules(r_id);
 		for(int i=0;i<s2.length;i++) {
 			integer= rolesModulesService.insertRoleModules(r_id, Integer.parseInt(s2[i]));
-			}
-		return Result.toClient(integer>0 ? true : false, integer>0 ? "设置成功" : "设置失败");
+		}
+		return Result.toClient(true, integer>0 ? "设置成功" : "设置失败");
 	}
 	
 	@RequestMapping(value="/selectByName",method=RequestMethod.POST)
@@ -78,4 +92,5 @@ public class RolesController {
 			return Result.toClient(false, "用户已经存在");
 		}
 	}
+	
 }
