+++ src/com/spz/controller/UsersController.java
@@ -10,10 +10,15 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.spz.entity.Roles;
+import com.spz.entity.Student;
 import com.spz.entity.UserRoles;
+import com.spz.entity.Userchecks;
 import com.spz.entity.Users;
+import com.spz.service.StudentService;
 import com.spz.service.UserRolesService;
+import com.spz.service.UserchecksService;
 import com.spz.service.UsersService;
+import com.spz.util.InitDateTime;
 import com.spz.util.Result;
 
 @Controller
@@ -22,6 +27,10 @@ public class UsersController {
 	@Autowired UsersService usersService;
 	
 	@Autowired UserRolesService userRolesService;
+	
+	@Autowired StudentService studentService;
+	
+	@Autowired UserchecksService userchecksService;
 	
 	@RequestMapping(value="/moduls",method=RequestMethod.POST)
 	@ResponseBody
@@ -38,7 +47,6 @@ public class UsersController {
 	@RequestMapping(value="/insertUser",method=RequestMethod.POST)
 	@ResponseBody
 	public String insertUser(Users users) {
-		System.out.println(users.toString());
 		Integer num=null;
 		if(users.getU_id()==null) {
 			num=usersService.insertUsers(users);
@@ -51,8 +59,15 @@ public class UsersController {
 	
 	@RequestMapping(value="/deleteUser",method=RequestMethod.POST)
 	@ResponseBody
-	public Integer deleteUser(Integer u_id) {
-		return usersService.deleteUsers(u_id);
+	public String deleteUser(Integer u_id) {
+		//删除员工前判断学生表是否有与之关联的学生还‘未缴费’的
+		List<Student> byU_id = studentService.selectStudentByU_id(u_id);
+		if(byU_id.size()>0) {
+			return Result.toClient(true, "该员工还有学生没有缴费！您不能删除！");
+		}else {
+			Integer users = usersService.deleteUsers(u_id);
+			return Result.toClient(true,users>0 ? "删除成功" : "删除失败" );
+		}
 	}
 	
 	@RequestMapping(value="/selectRoles",method=RequestMethod.POST)
@@ -93,6 +108,121 @@ public class UsersController {
 		}else{
 			return Result.toClient(false, "用户已经存在");
 		}
-		//return Result.toClient(userByName==null ? false : true, userByName == null ? "" : "用户已经存在");
 	}
+	
+	@RequestMapping(value="/selectUserIsFouDaKa",method=RequestMethod.POST)
+	@ResponseBody
+	public String selectUserIsFouDaKa(Integer u_id) {
+		if(u_id==null) {
+			return Result.toClient(true, "");
+		}
+		Userchecks userchecks = userchecksService.selectUserchecks(u_id);
+		//判断用户是否已经打卡了
+		if("是".equals(userchecks.getUs_checkState())) {
+			return Result.toClient(true, "1");
+		}else {
+			return Result.toClient(true, "0");
+		}
+	}
+	
+	@RequestMapping(value="/updataUserchecksIsDaka",method=RequestMethod.POST)
+	@ResponseBody
+	public Integer updataUserchecksIsDaka(Userchecks userchecks) {
+		userchecks.setUs_checkinTime(InitDateTime.initTime());
+		return userchecksService.updateUserchecks(userchecks);
+	}
+	
+	@RequestMapping(value="/selectUserByPwd",method=RequestMethod.POST)
+	@ResponseBody
+	public Integer selectUserByPwd(Users users) {
+		Users userBylogin = usersService.selectUserBylogin(users);
+		if(userBylogin!=null) {
+			return 1;
+		}else {
+			return 0;
+		}
+	}
+	
+	@RequestMapping(value="/updateUsersByPwd",method=RequestMethod.POST)
+	@ResponseBody
+	public String updateUsersByPwd(Users users) {
+		Integer updateUsers = usersService.updateUsers(users);
+		return Result.toClient(true, updateUsers>0 ? "修改成功" : "修改失败");
+	}
+	
+	@RequestMapping(value="/selectUsersByKaiQi",method=RequestMethod.POST)
+	@ResponseBody
+	public String selectUsersByKaiQi(Integer u_id) {
+		Users usersByKaiQi = usersService.selectUsersByKaiQi(u_id);
+		return Result.toClient(usersByKaiQi.getU_state() == 1 ? true : false, "");
+	}
+	
+	//经理点击开启分量的按钮和取消分量使用的
+	@RequestMapping(value="/updateUsersByu_state",method=RequestMethod.POST)
+	@ResponseBody
+	public Integer updateUsersByu_state(Integer u_state) {
+		return usersService.updateUsersByu_state(u_state);
+	}
+	
+	@RequestMapping(value="/selectUsersByflcz",method=RequestMethod.POST)
+	@ResponseBody
+	public String selectUsersByflcz(Users users){
+		return usersService.selectUsersByflczAll(users);
+	}
+	
+	@RequestMapping(value="/updateUserByIdState",method=RequestMethod.POST)
+	@ResponseBody
+	public Integer updateUserByIdState(Users users) {
+		return usersService.updateUsers(users);
+	}
+	
+	/*@RequestMapping(value="/updateUserByIdState",method=RequestMethod.POST)
+	@ResponseBody
+	public String selectUsersByfenliang() {
+		return usersService.selectUsersByflczs();
+	}*/
+	
+	
+	
+	
+	/*孙所蕾*/
+	@RequestMapping(value="/selectUserCheckAll",method=RequestMethod.POST)
+	@ResponseBody
+	public String selectUserCheckAll(Userchecks userchecks) {
+		
+		return userchecksService.selectUserCheckAll(userchecks);
+	}
+	
+	@RequestMapping(value="/updateUserchecksPL",method=RequestMethod.POST)
+	@ResponseBody
+	public Integer updateUserchecksPL(Userchecks userchecks,String usid) {
+		userchecks.setUs_checkinTime(InitDateTime.initTime());
+		String[] ssid=usid.split(",");
+		Integer ii =0;
+		for (int i = 0; i < ssid.length; i++) {
+			Integer sid = Integer.parseInt(ssid[i]);
+			userchecks.setUs_id(sid);
+			ii=userchecksService.updateUserchecksPL(userchecks);
+		}
+		return ii;
+	}
+	
+	@RequestMapping(value="/selectUsersByYuanGong")
+	@ResponseBody
+	public String selectUsersByYuanGong() {
+		return userchecksService.selectUsersByYuanGong();
+	}
+	
+	@RequestMapping(value="/selectUsersJingLi")
+	@ResponseBody
+	public String selectUsersJingLi() {
+		return usersService.selectUsersJingLi();
+	}
+	
+	@RequestMapping(value="/updateUserchecks")
+	@ResponseBody
+	public Integer updateUserchecks(Userchecks userchecks) {
+		return userchecksService.updateUserchecksPL(userchecks);
+	}
+	
 }
